Complete ZcashRseed type
* by moving to its proper folder.
* marking as complete in STATUS.md.
* Adding comments.
