+++ lib/uniffi-zcash/src/zcash_primitives/sapling/note/mod.rs
@@ -2,12 +2,12 @@ mod commitment;
 pub use self::commitment::*;
 
 use crate::{
-    ZcashPaymentAddress, ZcashResult, ZcashRseed,
+    utils::cast_slice, ZcashError, ZcashJubjubFr, ZcashPaymentAddress, ZcashResult,
     ZcashSaplingNoteValue,
 };
 
 use std::sync::Arc;
-use zcash_primitives::sapling::Note;
+use zcash_primitives::sapling::{Note, Rseed};
 
 pub struct ZcashSaplingNote(Note);
 
@@ -46,5 +46,37 @@ impl ZcashSaplingNote {
 impl From<Note> for ZcashSaplingNote {
     fn from(inner: Note) -> Self {
         ZcashSaplingNote(inner)
+    }
+}
+
+/// Enum for note randomness before and after [ZIP 212](https://zips.z.cash/zip-0212).
+///
+/// Before ZIP 212, the note commitment trapdoor `rcm` must be a scalar value.
+/// After ZIP 212, the note randomness `rseed` is a 32-byte sequence, used to derive
+/// both the note commitment trapdoor `rcm` and the ephemeral private key `esk`.
+pub enum ZcashRseed {
+    /// This expects the data from ZcashJubjubFr,
+    /// which can be obtained by calling ZcashJubjubFr::to_bytes().
+    BeforeZip212 {
+        fr_data: Vec<u8>,
+    },
+    AfterZip212 {
+        data: Vec<u8>,
+    },
+}
+
+impl TryFrom<ZcashRseed> for Rseed {
+    type Error = ZcashError;
+
+    fn try_from(value: ZcashRseed) -> Result<Self, Self::Error> {
+        match value {
+            ZcashRseed::BeforeZip212 { fr_data } => Ok(Rseed::BeforeZip212(
+                ZcashJubjubFr::from_bytes(fr_data.as_slice())?.into(),
+            )),
+            ZcashRseed::AfterZip212 { data } => {
+                let casted_data = cast_slice(data.as_slice())?;
+                Ok(Rseed::AfterZip212(casted_data))
+            }
+        }
     }
 }
